[Chore #161995712] Separate tests into different files
